updated save data, discovered bug in agent genomes -__-
Discovered what appears to be a pretty big bug in the genetic algorithm.
I decided to have the simulation keep track of a random agent (I've
called him Henry) for the duration, and in the course of updating the
way data is saved in the stats file I discovered that, at least in
Henry's case, his genome is always empty of weights. Since the Henry is
chosen at random when the simulation is created, I can generalize this
to say that none of the agents have genomes -__-. This needs to be
addressed.