may have fixed the bug
I wasn't overriding the default #num_weights method in the Agency
module, which just returned zero. But in the simulation constructor, I
rely on this return value to create the genomes for the agents XD. I'm a
bit skeptical that this has solved the problem completely, because the
SeekingNet isn't fully constructed until its first response, and then
the hidden layer changes with every subsequent response. But perhaps
this will work out fine since the changes happen consistently. We'll
see.