modified Henry, fixed issues with NaNs in network
I was returning Float::MAX as a signal from sensors that were receiving
no input, but what I didn't anticipate was that once you do any sort of
arithmetic to that number that makes it  bigger, it changes to Infinity,
which doesn't play nicely with arithmetic in general. So I just replaced
it with like a million or something, some ridiculously high number that
is never going to be within range of an agent.

Also, I created a sprite for our dear Henry, so we can easily identify
him in the simulation. It's subtle enough that it doesn't keep your
attention once its been drawn, so it doesn't interfere with observing
the simulation as a whole.